Add include path of FreeType 2 as FTGL does not pull it in automatically
Apparently FTGL's CMake find module does not pull in the include path for
FreeType 2 headers automatically. So we need to do that manually on our
own. This code is relying on pkg-config for finding the include path. As
pkg-config has been intentionally avoided for Windows builds I only enabled
this for mingw-w64 builds so far (where pkg-config is usually available).